Remove simulate button from main windows
But software still supports use "A" or "Insert" key to simulate a new sensor added to the system, use "D" or "Delete" key to simulate an existed sensor was removed from the system. Because so far we only have one sensor in hand, but our software supports multiple sensors are added or removed from system dynamically.