Completed MongoMVC in another Tab, however...
Second seperate window opens along side it because I'm still inovking
the MongoView class in buildMongoPanel method, but at least it's content
stay within the mongoViewPanel JPanel for the tab, while invoking the
Mongo model/control for it to work as intended. This the best I can
accomplish with no hiccups in the code, otherwise I'd have to
restructure the SonarPingView class unnecessarily.